refactor(floatTicket): change import dirs
