+++ src/features/tickets/components/floatTicket/FloatTicketSkeleton.tsx
@@ -3,7 +3,7 @@ import Spacer from '../../../../components/spacer/Spacer'
 import FloatTicketHeaderSkeleton from '../floatTicketHeader/FloatTicketHeaderSkeleton'
 import StatusBarSkeleton from '../statusBar/StatusBarSkeleton'
 import CommentSkeleton from '../comment/CommentSkeleton'
-import CreateCommentFormSkeleton from '../../forms/create-comment-form/CreateCommentFormSkeleton'
+import CreateCommentFormSkeleton from '../../forms/createCommentForm/CreateCommentFormSkeleton'
 import Style from './FloatTicet.module.css'
 import FloatTicketAuthorSkeleton from '../floatTicketAuthor/FloatTicketAuthorSkeleton'
 import FloatTicketInfoSkeleton from '../floatTicketInfo/FloatTicketInfoSkeleton'
